🚚 move user identity repository to module
