+++ backend/modules/user/userIdentity/repository/get_user_identity.py
@@ -3,7 +3,7 @@ from uuid import UUID
 
 from models import get_supabase_client
 from modules.user.userIdentity.entity.user_identity import UserIdentity
-from repository.user_identity.create_user_identity import create_user_identity
+from modules.user.userIdentity.repository import create_user_identity
 
 logger = get_logger()
