+++ backend/modules/user/userIdentity/repository/update_user_properties.py
@@ -3,8 +3,8 @@ from uuid import UUID
 
 from models.settings import get_supabase_client
 from modules.user.userIdentity.entity.user_identity import UserIdentity
+from modules.user.userIdentity.repository import create_user_identity
 from pydantic import BaseModel
-from repository.user_identity.create_user_identity import create_user_identity
 
 
 class UserUpdatableProperties(BaseModel):
